feat(notifications): Phase 200 系统通知标记已读 API (POST /api/v1/system/notifications/{id}/mark-read)
- 新增 POST 端点支持标记个人通知为已读
- JWT 认证，登录用户可访问
- 权限验证：普通用户只能标记自己的通知，admin 可标记任意
- 404 Not Found (通知不存在)
- 403 Forbidden (无权标记)
- 409 Conflict (已是已读状态)
- 返回 200 OK + 通知摘要信息
- 使用 SqliteNotificationRepository 实现真实数据库操作
- 添加 docs/system_notifications_mark_read_api.md 文档
- 清理废弃文件 system_notifications_mark_read_post.rs

+++ src/handlers/system_notifications_mark_read.rs
@@ -175,7 +175,7 @@ pub async fn mark_notification_as_read(
             if !is_admin {
                 // 非 admin 用户只能标记自己的通知或系统通知
                 if let Some(target_id) = notification.target_user_id {
-                    if target_id != user_id {
+                    if target_id != user_id as i64 {
                         return Ok(HttpResponse::Forbidden().json(ErrorResponse {
                             success: false,
                             error: "只能标记自己的通知为已读".to_string(),
